requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33
python-dotenv==1.0.0
pyyaml==6.0.1
kiteconnect==4.2.0

# Technical analysis
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Visualization
matplotlib==3.8.2
plotly==5.18.0
streamlit==1.29.0

# API clients (optional - uncomment as needed)
# alpaca-trade-api==3.0.2
# ccxt==4.1.90  # For cryptocurrency exchanges

# Database (optional)
# sqlalchemy==2.0.23

# Utilities
requests==2.31.0
pytz==2023.3
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
